CDRIVER-1571 simplify async command loop
mongoc_async_run() no longer returns "true" if there is more work to do:
it never returns early now since there is only one timeout for the whole
function. Also, make timeouts int64_t in all internal functions.

+++ src/mongoc/mongoc-async-cmd-private.h
@@ -80,7 +80,7 @@ mongoc_async_cmd_new (mongoc_async_t           *async,
                       const bson_t             *cmd,
                       mongoc_async_cmd_cb_t     cb,
                       void                     *cb_data,
-                      int32_t                   timeout_msec);
+                      int64_t                   timeout_msec);
 
 void
 mongoc_async_cmd_destroy (mongoc_async_cmd_t *acmd);
